rename utils remove BatchOperations refactor DocxRemoveBookMark
