+++ src/main/java/com/ewp/crm/models/Client.java
@@ -1,7 +1,7 @@
 package com.ewp.crm.models;
 
 import com.ewp.crm.models.whatsapp.WhatsappMessage;
-import com.ewp.crm.utils.patterns.ValidationPattern;
+import com.ewp.crm.util.patterns.ValidationPattern;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.apache.commons.lang3.builder.DiffBuilder;
 import org.apache.commons.lang3.builder.DiffResult;
