+++ src/main/java/com/ewp/crm/models/User.java
@@ -1,6 +1,6 @@
 package com.ewp.crm.models;
 
-import com.ewp.crm.utils.patterns.ValidationPattern;
+import com.ewp.crm.util.patterns.ValidationPattern;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.hibernate.annotations.Fetch;
